+++ idea/src/org/jetbrains/kotlin/idea/injection/KotlinLanguageInjector.kt
@@ -16,6 +16,7 @@
 
 package org.jetbrains.kotlin.idea.injection
 
+import com.intellij.codeInsight.AnnotationUtil
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
@@ -27,6 +28,8 @@ import org.intellij.plugins.intelliLang.Configuration
 import org.intellij.plugins.intelliLang.inject.InjectorUtils
 import org.intellij.plugins.intelliLang.inject.config.BaseInjection
 import org.intellij.plugins.intelliLang.inject.java.JavaLanguageInjectionSupport
+import org.jetbrains.kotlin.idea.util.findAnnotation
+import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.*
 import java.util.*
 
@@ -140,7 +143,14 @@ class KotlinLanguageInjector : LanguageInjector {
         val argumentIndex = (argument.parent as KtValueArgumentList).arguments.indexOf(argument)
         val ktParameter = ktFunction.valueParameters.getOrNull(argumentIndex) ?: return null
 
-        return findInjection(ktParameter, Configuration.getInstance().getInjections(KOTLIN_SUPPORT_ID))
+        val patternInjection = findInjection(ktParameter, Configuration.getInstance().getInjections(KOTLIN_SUPPORT_ID))
+        if (patternInjection != null) {
+            return patternInjection
+        }
+
+        val injectAnnotation = ktParameter.findAnnotation(FqName(AnnotationUtil.LANGUAGE)) ?: return null
+        val languageId = extractLanguageFromInjectAnnotation(injectAnnotation) ?: return null
+        return InjectionInfo(languageId, null, null)
     }
 
     private fun findInjection(element: PsiElement?, injections: List<BaseInjection>): InjectionInfo? {